samples: conn_time_sync: Adapt timed toggle for RTC overflows
This commit adapts the timed toggle implementation to not toggle
upon RTC overflows similar to how the iso_time_sync sample was adapted.
This would not occur in the conn_time_sync sample as that sample
toggled every N ms, thus never reaching the RTC overflow.

The code is adapted to maintain consistency between the two
implementations. Future work is to extract out this to a common
module.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>