samples: bluetooth: llpm: Disable BT debug logging
When both: CONFIG_LOG, CONFIG_BT_LOG_LEVEL_DBG are enabled the BT
Host prints plenty of data on UART. That makes the llpm sample
application nonfunctional. It looks like the debug logs printing
takes so much time that it is hard to establish a connection.
In case of luck and link is established then security level change
doesn't happen, the security_changed callback is never called.

To make the sample functional, the debug logs are disabled.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>