nrf_security: Refactor Cracen IKG keys
This refactors how we handle the IKG key IDs for
Cracen. Before this change we used the internal
Cracen IKG key identifiers inside the builtin key
driver. This had an issue because both the KMU keys
and the IKG keys had identifiers in the slots 0-2.

To avoid this collision the IKG handling is refactor
to use the reserved PSA key identifiers in the driver
level and only use the internal key intentifiers deeper
in the implementation in order to avoid the conflicts.

This also removes unused structs related to the IKG
keys and unused code as well.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>