mgmt: suitfu: Make main image inactive in recovery
Until now even when in recovery firmware, the image
group marked the main application image as active.

This caused that the mobile application was not able
to recover the device - the confirm button for the
new application image was greyed out.

This commit marks the main application as inactive
when running the recovery firmware.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>