nrf_security: Always enable builtin keys for 54L
Enable the builtin keys for the nRF54L series when
nrf_security is being used. This option for
builtin keys is needed for the KMU keys which
don't always have dependencies on the hardware unique key
or the identity key.

We guard this option on the PSA Cracen driver because the
driver implements the necessary builtin key functions.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>