samples: peripheral: lpuart: Cleanup in debug configuration
Simplify debug configuration. Remove debug conf and overlay file.
All targets are now configuring console UART to be in the lowest
power state when not transmitting logs. With this configuration it
is possible to have console uart enabled in debug and none debug
build. It is also possible to reach the lowest energy state in
the debug build, too.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>