nrf_security: Refactor mutexes
This refactors the mutexes in nrf_security.

The main issue with the current design is that
that it doesn't use the Zephyr macro K_MUTEX_DEFINE
to define the mutexes. If this macro is not used
the Zephyr mutexes need to be initialized at runtime,
but the Cracen code doesn't initalize them and so
they don't work properly.

This simplifies the nrf_security mutexes by removing
the flag. This keep the design simpler and less
error prone. An extra flag added to the Zephyr mutexes
doesn't seem to be necessary. The nrf_security mutexes
were added in order to be used by the Cracen driver and
reduce the ifdef conditions inside the driver code
for tfm/non tfm builds. No initialization is required
for these mutexes because there are always meant to be
statically defined and this removes the need for
initialization as long as we use the Zephyr macros
for this.

The mbedtls prefixed mutexes used by the PSA core are now
changed to direcly use the zephyr mutexes. This makes the
design simpler again, since we remove the connection with the
nrf_security mutexes. The mbedlts prefixed mutexes are always
under ifdef conditions inside the PSA core. So using the mechanism
of the nrf_security mutexes doesn't improve anything.

Since both the nrf_security mutexes and the mbedtls prefixed
mutexes indide PSA core are basically just a front end which
use the Zephyr mutexes its better to keep them separated.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>