samples: protocols_serialization: use different GPIO pins of UART
One of the previously selected pins for uart21 peripheral
was also used to drive LED 1 on nRF54L15 DK, which might
have interferred with the UART hardware flow control.

Due to lack of available pins on GPIO port 1, use the pins
dedicated to Button 1 & 2 but make sure that the buttons
are removed from the devicetree to prevent from using them
accidentally.

Also, fix pull-up configuration for UART pins.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>