net: openthread: rpc: don't bring interface up in IF_ENABLE command
The OpenThread RPC server starts the network interface
automatically on boot because otherwise no frames are
delivered to the OpenThread stack.

Therefore, don't bring the interface up or down when
IF_ENABLE RPC command is received. Instead, assume that
IF_ENABLE command only indicates that the client has
brought its interface up or down, and that it is
interested (or not) in receiving all incoming packets.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>