samples: wifi: radio_test: Fix crash during temp and rssi get
Get temperature and RSSI commands don't take any args, but due to
copy-paste, the code tries to parse the args, argv[1] without any NULL
or boundary checks causing a crash.

This bug is quite old but in nRF53 (and 52) the address 0x0 is a valid
address (IIRC mapped to begining of external flash), so, the conversion
to string passes, though it returns garbage but the check for the
returned value is completely unnecessary, so it doesn't matter if that
is invoked or not, because of this, the issue was hidden.

In nRF54H the address 0x0 is not mapped and is considered invalid and
leading to the crash.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>