Bluetooth: Tester: remove unnecessary lightness server from composition
Removing the lightness server from the composition. The reasons for
doing so; PTS can not chose the correct element to run tests when there
are multiple lightness server instances on different elements. There is
an element already with a lightness server along with xyL and hsL
servers, this element can be used to run all lightness server tests.

Signed-off-by: alperen sener <alperen.sener@nordicsemi.no>