mpsl: init: Make the CONFIG_MPSL_HFCLK_LATENCY user configurable
The MPSL uses the CONFIG_MPSL_HFCLK_LATENCY Kconfig option to setup
HFCLK latency. There is an mpsl_clock_hfclk_latency_set() API that
a user can call to setup the HFCLK latency. It can also be done
automatically on a user behalf while mpsl_init() is executed.
That will be possible when CONFIG_MPSL_HFCLK_LATENCY is user
configurable.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>