nrf_security: Fix dependency for threading_alt.h
This fixes a dependency when including the threading_alt.h
from nrf_security.

The CC3XX platform library provides the implementation of the
mbedtls_mutex functions and it expects the mbedTLS mutexes to be of
type nrf_cc3xx_platform_mutex_t.
The definitions of the mutexes included in the files
nrf_cc3xx_platform_mutex_(zephyr/freertos).c

The threading_alt.c in nrf_security defines the mbedTLS mutexes
as Zephyr mutexes (k_mutex). This doesn't work when the CC3XX
platform library is used, so here this dependency is added.

Ref: NCSDK-31155

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>