tests: benchmarks: power_consumption: Allow controlled thread suspension
Test was suspending test thread from the timer interrupt at random
(but periodic) point. Some drivers which are tested in that framework
handles power management in the thread context. Randomly suspending
the test thread could lead to a case where driver did not finish
requested operation and system goes to s2ram in the middle of the
driver operation. Extending the test framework to allow more
controlled thread suspension. Timer can indicate to the thread that
is shall be suspended and thread calls self suspension when ready.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>