tests: benchmarks: power_consumption: flash: fix exmif pm
The exmif instance used for the external flash operation
needs pm device runtime to be enabled for it, without it
the bus can not suspend, and s2ram can not be entered.

Additionally, the test needs to omplement suspend_req to
ensure the flash device and bus are suspended before the
thread is suspended.

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>