Bluetooth: Mesh: Fix idx in dtt_srv_get
Finds the element index used in `bt_mesh_dtt_srv_get` based on the
composition data instead of the runtime address.

Previously, the index used to locate the DTT Server was computed based
on the runtime address set during provisioning. As the Scene Setup
Server extends this model (which isn't necessarily present in the same
element), model initialization would fail as the address wasn't yet set.

Signed-off-by: Håvard Reierstad <haavard.reierstad@nordicsemi.no>