net: lib: downloader: Allow redirect to HTTPS from HTTP
Separate URI parsing and protocol detection to its own
function and when receiving the "Location: <URI>" header,
parse the URI again, so the next connection will use the
correct protocol and port.

Do not convert all HTTP headers to lowercase as URI
is case-sensitive. Convert just the header names, not values
after the first colon.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>