bluetooth: services: wifi_prov: Fix setting of nanopb_BUILD_RUNTIME
`nanopb_BUILD_RUNTIME` is not an option to be passed to nanopb,
but rather a cmake one, so set it to off in the proper way.

The previous invalid form was not working as intended and only by
chance not causing any problems. With a correction that was applied
in the nanopb module (removal of leading space before the first -I
directive passed to nanopb), this behavior changed and because of
that the generation could not be performed correctly.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>