nrf_security: cracen: kmu: Support KMU reserved area through DTS
When execution in place (CONFIG_XIP) is not enabled, which in practice
means that when Zephyr is built for a RAM loaded image, the Zephyr
linker script always places the RAM loaded image in the top address
of the RAM and then loads the linker scripts defined with the Zephyr
SECTION_PROLOGUE macros.

SECTION_PROLOGUE Zephyr macros was used to set the address of the
kmu_push_area making it incompatible with RAM loaded images.

The Zephyr reserved-memory devicetree methodology works for both use
cases but it requires heavy updates of multiple device tree files and
overlays. In order to support the RAM loaded images use cases faster
initial support for reserving the memory of nrf_kmu_reserved_push_area
though devicetree is limited to RAM loaded images.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>