mpsl: clock_ctrl: Fix wait for LFCLK if timeout returned earlier
There is a workaround implemented in the MPSL clock control
integration layer for cases where nrf2 clock control doesn't
get response from sysctrl for the LFCLK request.

The workaround was incorrectly implemented. After clock driver
returned -NRF_ETIMEDOUT new call to m_lfclk_wait() was done.
The new call waits on a semaphore that also returns timeout error.
The second m_lfclk_wait() call returns -NRF_EFAULT error code that
causes an assert in later stage of initialization.

The change done is to make sure if a clock driver returns
-NRF_ETIMEDOUT error for m_lfclk_wait then next m_lfclk_wait call
immediately returns 0 (as if the clock was ready). That is based
on an assumption:
- the sysctrl implicitly provides a LFCLK with accuracy that is
  the same or better than requested by MPSL,
- there are not other LFLCK requests from Radio core FW
- if there is another LFCLK request put, it is for a LFCLK with
  accuracy the same or better than required by radio protocols.

The change also affects LFCLK request and response APIs, so that
the new requests or released for the clock will return immediately
until the mpsl_clock_ctrl_uninit() is executed, that resets the sate
of the m_lflclk_req_timeout flag.

There is added a KConfig option: CONFIG_MPSL_EXT_CLK_CTRL_LFCLK_REQ-
_TIMEOUT_ALLOW that can be used to disable the workaround.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>